Rename CardContainer to MainView
