+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import fetch from 'node-fetch';
-import CardContainer from './components/CardContainer.js'
+import MainView from './components/MainView.js'
 import Navbar from './components/Navbar'
 // import './App.css';
 
@@ -75,7 +75,7 @@ export default class App extends Component {
 
         <h1>Welcome to Blue Ocean!</h1>
         <Navbar />
-        <CardContainer events={this.state.eventsPlaceHolder}/>
+        <MainView events={this.state.eventsPlaceHolder}/>
       </>
     );
   }
